tests/wpt/web-platform-tests/service-workers/service-worker/skip-waiting-using-registration.https.html

<!DOCTYPE html>
<title>Service Worker: Skip waiting using registration</title>
<script src="/resources/testharness.js"></script>
<script src="resources/testharness-helpers.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>

promise_test(function(t) {
    var scope = 'resources/blank.html';
    var url1 = 'resources/empty.js';
    var url2 = 'resources/skip-waiting-worker.js';
    var frame, frame_sw, sw_registration, oncontrollerchanged;
    var saw_controllerchanged = new Promise(function(resolve) {
        oncontrollerchanged = function(e) {
            assert_equals(e.type, 'controllerchange',
                          'Event name should be "controllerchange"');
            assert_true(
                e.target instanceof frame.contentWindow.ServiceWorkerContainer,
                'Event target should be a ServiceWorkerContainer');
            assert_equals(e.target.controller.state, 'activating',
                          'Controller state should be activating');
            assert_equals(
                frame_sw.controller.scriptURL, normalizeURL(url2),
                'Controller scriptURL should change to the second one');
            resolve();
        };
      });
    return service_worker_unregister_and_register(t, url1, scope)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(f) {
          frame = f;
          frame_sw = f.contentWindow.navigator.serviceWorker;
          assert_equals(
              frame_sw.controller.scriptURL, normalizeURL(url1),
              'Document controller scriptURL should equal to the first one');
          frame_sw.oncontrollerchange = t.step_func(oncontrollerchanged);
          return navigator.serviceWorker.register(url2, {scope: scope});
        })
      .then(function(registration) {
          sw_registration = registration;
          add_completion_callback(function() {
              frame.remove();
              registration.unregister();
            });
          return saw_controllerchanged;
        })
      .then(function() {
          assert_not_equals(sw_registration.active, null,
                            'Registration active worker should not be null');
          fetch_tests_from_worker(sw_registration.active);
        });
  }, 'Test skipWaiting while a client is using the registration');

</script>
